Refactor event set methods from account for reuse with works
